+++ packages/doc-generator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openpkg-ts/doc-generator
 
+## 0.6.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @openpkg-ts/ui@0.1.2
+
 ## 0.6.6
 
 ### Patch Changes